edit order

+++ src/infrastructure/http/routing/ordersRouting.ts
@@ -1,10 +1,17 @@
 import {FastifyInstance, FastifyReply, FastifyRequest} from "fastify";
-import {createOrderSchema, deleteOrderSchema, getOrderSchema, getOrdersSchema} from "../schemas/orderSchema";
 import {
-    createOrderController, deleteOrderController,
+    createOrderSchema,
+    deleteOrderSchema,
+    editOrderSchema,
+    getOrderSchema,
+    getOrdersSchema
+} from "../schemas/orderSchema";
+import {
+    createOrderController, deleteOrderController, editOrderController,
     getAllOrderController,
     getByIdOrderController
 } from "../../../application/controllers/orderController";
+import {Roles} from "../../../domain/user/valueObjects/role";
 
 
 export function registerOrderRouting(fastify: FastifyInstance) {
@@ -16,12 +23,20 @@ export function registerOrderRouting(fastify: FastifyInstance) {
         await getByIdOrderController(req, res)
         // получить заказ по id
     });
+    fastify.put('/order/:id',editOrderSchema, async (req: FastifyRequest<OrderRequest>, res: FastifyReply) => {
+        await req.jwtVerify()
+        //@ts-ignore
+        if(req.user.data.role != Roles.admin) return res.status(403).send('Not authorized')
+        await editOrderController(req, res)
+    });
     fastify.post('/order',createOrderSchema, async (req: FastifyRequest<OrderRequest>, res: FastifyReply) => {
         await createOrderController(req, res)
         // создать заказ
     });
     fastify.delete('/order/:id', deleteOrderSchema, async (req: FastifyRequest<OrderRequest>, res: FastifyReply) => {
+        await req.jwtVerify()
+        //@ts-ignore
+        if(req.user.data.role != Roles.admin) return res.status(403).send('Not authorized')
         await deleteOrderController(req, res)
-        // удалить заказ
     });
 }
